requirements.txt
Django
psycopg2
requests
djangorestframework
pyrebase
pillow
django-cors-headers
python-dotenv
